examples/steel-playwright-python-starter/requirements.txt
playwright==1.48.0
python-dotenv==1.0.0
steel-sdk==0.1.0b1